requirements.txt
flet==0.24.1
requests==2.32.3
mysql-connector-python
plyer==2.1.0
bcrypt==4.2.0
reportlab==4.2.5
screeninfo
Pillow==10.4.0
pyyaml==6.0.1
python-dotenv==1.0.0
